+++ src/ActivationFunction.cpp
@@ -1,6 +1,6 @@
 #include "..\include/ActivationFunction.h"
 
-double ActivationFunction::ReLU(double value){
+float ActivationFunction::ReLU(float value){
     if (value < 0) {return 0;}
     return value;
 }
